Added removal, fixed alteration and added untested shrink.

+++ cleanFiles.sh
@@ -1,4 +1,13 @@
-rm dados.txt
-rm primario.ndx
-rm primario.lock
+if [ -f "dados.txt" ]
+then
+    rm dados.txt
+fi
+if [ -f "primario.ndx" ]
+then
+    rm primario.ndx
+fi
+if [ -f "primario.lock" ]
+then
+    rm primario.lock
+fi
